+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 
-## 1.2.0+01
+## 2.0.0
+24 March 2022
+- Both Mixins on State<T>  Target only State objects.
+- Removed method, initFutureBuilderState()
+- Removed state from method, initInheritedState()
+
+## 1.2.0+02
 21 March 2022
 - Introduced the class, InheritController
-- Widget buildWidget(BuildContext context);
 
 ## 1.1.0
 20 March 2022